Major structural improvements to choose()
Code is now much cleaner, fully commented, and will raise exceptions in
edge cases.

Minor adjustments to the function call required adjusting every time the
method is called; now the dictionary is passed first, followed by
(optional) keys to look at inside it. This should be a net improvement.

Modified the dictionary to ensure that "no" is read as a string instead
of a boolean.